// src/containers/NonPOPaymentContainer/shared/enums.ts
export enum PMT_PROJECT_LINE_ITEM_KEY {
  LINE_NUMBER = 'lineNumber',
  PROJECT_NUMBER = 'projectNumber',
  SUB_PROJECT = 'subProject',
  BUDGET_CATEGORY = 'budgetCategory',
  SUB_BUDGET_CATEGORY = 'subBudgetCategory',
  SERVICE_DATE = 'serviceDate',
  AMOUNT = 'amount',
  DESCRIPTION = 'description',
}